[bugfixes-backport]VMware: enable a resize of instance with no root disk
This addresses the issue when the user wishes to update the CPU
or memory of the instance.

No changes to disks will be done if there is no root disk.

Examples of when there is no root disk:
1. When an instance is booted from a cinder volume.
2. When booting from an ISO and the root disk size is zero

In these cases there will not be a disk created on the vmware
shared datastore.

Closes-bug: #1444446
(cherry picked from commit 54afe44c7b68902556850baf4eba0846b000aa35)

Change-Id: Id14d8b94682b95fa3556126dc8b282f9c1038489
Signed-off-by: zhangjianfeng <kill51216@163.com>