[bugfixes-backport]Fix multipath iSCSI encrypted volume attach failure
Currently iSCSI volume attachment fails if iscsi_use_multipath is
set to True. This is because the encryptor requests cryptsetup
to create the symlink to the LUKS device with the same name of
the device-mapper multipath device. To avoid the name collision,
this patch adds the 'crypt-' prefix to the symlink.

NOTE(lyarwood): The rootwrap.d portion of this change is dropped for
stable/mitaka to avoid upgrade issues. The current wildcard based ln
filter means that the remainder of the change will still correct this
issue.

Conflicts:
        etc/nova/rootwrap.d/compute.filters
        nova/tests/unit/volume/encryptors/test_cryptsetup.py

Closes-Bug: #1439869
(cherry picked from commit 89a61ab8f4602e018763afb173e1a862f151a222)

Change-Id: I433dfb3a988324611d1efb9b43f448e1d76be177
Signed-off-by: zhangjianfeng <kill51216@163.com>