[bugfixes-backport]db: retry on deadlocks while adding an instance
We are hitting deadlocks in the gate when we are inserting the new
instance_extra row into the DB.

We should follow up this fix and look at way to avoid the deadlock
happening rather than retrying it. It currently doesn't happen too
often, so this should be enough to stop the problem while we work on a
better fix.

Closes-Bug: #1480305

(cherry picked from commit df15e467b61fee781e78b07bf910d6b411bafd44)

Change-Id: Iac3ca813029d81798e2082eaaa5103d83e3f6aac
Signed-off-by: zhangjianfeng <kill51216@163.com>