[bugfixes-backport]claims: Do not assume image-meta is a dict
MoveClaim is expected to accept the image_meta argument to it's __init__
which we assumed was always a dict. This is not the case anymore as we
pass in the ImageMeta object, that recently lost it's dict-mimicing
powers so the assumption that it's safe to call from_dict on it is no
longer true.

This patch make sure that we convert to objects internally always.

(cherry picked from commit b01187eede3881f72addd997c8fd763ddbc137fc)

Change-Id: Ieb0cdb64fa49399c7d3e9e187692531b64973c67
Signed-off-by: zhangjianfeng <kill51216@163.com>