[bugfixes-backport]Fix TestNeutronv2.test_deallocate_for_instance_2* race failures
The unit tests
    test_deallocate_for_instance_2_with_requested
    test_deallocate_for_instance_2
in
    nova.tests.unit.network.test_neutronv2.TestNeutronv2
fail randomly. An example error looks like this:

    b'mox3.mox.UnexpectedMethodCallError: Unexpected method call.
    unexpected:- expected:+'
    b"- Client.delete_port('my_portid1') -> None"
    b'? ^'
    b"+ Client.delete_port('my_portid2') -> None"
    b'?

This is due to the mox record with expected the deletion of multiple
ports on an instance in a specific order. This order is not important
though and also not the same in each test case execution. To make it
explicit in the test case, that the order is *not* important, I've
chosen to define that in the mock records with "InAnyOrder()".

Another solution could have been to ensure the order of the port IDs
through the unit tests, but that could be interpreted as a precondition
for the functional code to work correctly, which is a wrong assumption
and therefore not chosen for this fix.

Closes-Bug: 1521599

Change-Id: I089dcd220ee768bac6ec25accf8ec8eb925fe506
Signed-off-by: zhangjianfeng <kill51216@163.com>