[bugfixes-backport]Fix service version to update the DB
When new code is installed with a newer service version, the version in
the DB should get updated.

The fix here is to:
1) eliminate the special case for 'version' in save()
2) cause version to be saved on startup

NOTE(lyarwood): Single conflict caused by the use of Mock 1.3.0 on
stable/mitaka and as a result the lack of assert_called_once().

Conflicts:
        nova/tests/unit/test_service.py

Closes-Bug: 1579839
(cherry picked from commit 854c39e26d94f006f539ea993ffaafa4892b100c)

Change-Id: Ic7da45f3e0aa0fbe40c0f444fee7d5aaf1edfede
Signed-off-by: zhangjianfeng <kill51216@163.com>