[bugfixes-backport]Completed migrations are not "in progress"
This fixes an issue where we consider migrations that have been completed
as "in progress" by not filtering them out of the result set from
migration_get_in_progress_by_host_and_node(). This adds that state to
the list and adds a migration in that state to the test case. If it
is not filtered, the counts won't line up and the test will fail.

Closes-Bug: #1583720
(cherry picked from commit e5b8060c08bb972a1960575749f240d7160bc114)

Change-Id: Ia44583a801f1b96d1bcc1398b8934a53cfacf9e6
Signed-off-by: zhangjianfeng <kill51216@163.com>