[bugfixes-backport]Ironic nodes with instance_uuid are not available
Currently, if a node is in AVAILABLE or NOSTATE (legacy), regardless of
if it has an instance_uuid it is considered able to be scheduled to.
However, it's impossible for a deployment to succeed to an ironic node
with instance_uuid populated. We should not schedule to nodes in this
state.

Closes-bug: 1503453
(cherry picked from commit ffe3093786b137f46646c4cf5fb3ba6b597418a9)

Change-Id: I97d1d7a5c520dd4b63c29cb7c9cf307ede8384cd
Signed-off-by: zhangjianfeng <kill51216@163.com>