[bugfixes-backport]Handle keypair not found from metadata server
With commit e83842b80b73c451f78a4bb9e7bd5dfcebdefcab we
attempt to load keypairs for an instance from instance_extra,
but if that hasn't been migrated yet we fall back to loading
the keypair from the database by name.

If the keypair was deleted, the instance object will just set
an empty KeyPairList for instance.keypairs and we'll get an
IndexError when using self.instance.keypairs[0] in
_metadata_as_json.

This adds a check that instance.keypairs actually has
something in it. If not, we log a message and don't return
any key values in the metadata dict - same as if instance.key_name
wasn't set to begin with.

Conflicts:
        nova/api/metadata/base.py

NOTE(mriedem): This is basically a different patch given in mitaka
the keypairs weren't migrated to the instance_extra table, so this
is more or less a mitaka-specific fix for the same KeypairNotFound
failure.

Closes-Bug: #1592167
(cherry picked from commit 4317166b72bb0aadd0321acdf9f2450c1a99d0a4)

Change-Id: Ifa354bbbffebc1ce3a7b21a3505cf8088c06c27a
Signed-off-by: zhangjianfeng <kill51216@163.com>