[bugfixes-backport]Do not try to backport when db has older object version
Instance extras are stored as serialized objects in the database and
because of this it is possible to get a version back that is older
than what the client requested. This is in itself not a problem, but
the way we do things right now in conductor we end up trying to
backport to a newer version, which raises InvalidTargetVersion.

This change adds a check to make sure we skip backporting if the
requested version is newer than the actual db version as long as the
major version is the same.

Closes-Bug: #1596119
(cherry picked from commit 1e3e7309997f90fbd0291c05cc859dd9ac0ef161)

Change-Id: Ic404a937d183690b3a26da76224cde644ce3ff8c
Signed-off-by: zhangjianfeng <kill51216@163.com>