[bugfixes-backport]Return HTTP 200 on list for invalid status
The server listing API raises a 500 error if you pass an invalid status
filter for admin user. In the case of a non-admin user, it simply
returns an empty list. In the case of an admin user, it fetches extended
server attributes, so a condition was added to get extended server
attributes only when servers list is not empty.

This change simply removes the cause of the 500 exception. A subsequent
patch with a microversion bump will modify the behavior so that a 400
Bad Request will be raised for an invalid status, for both admin and
non-admin alike.

Co-Authored-By: Dinesh Bhor <dinesh.bhor@nttdata.com>

Closes-Bug: #1579706

(cherry picked from commit ee4d69e28dfb3d4764186d0c0212d53c99bda3ca)

Change-Id: Id9503615102a84e8094f70fda5b72021ce484e33
Signed-off-by: zhangjianfeng <kill51216@163.com>