[bugfixes-backport]rbd_utils: wrap blocking calls in tpool.Proxy()
librbd is a Python binding around a C library, which is not aware of
eventlet - all the calls to the functions from this library will block
the whole nova-compute process for duration of a call. To make sure
nova-compute remains responsive we need to wrap all the calls in
tpool.Proxy() eventlet helper, that switches the execution context
back to the event loop, while the call is executed in a native OS
thread from a pool.

Prefer tpool.Proxy() to tpool.execute() here as the former allows for
wrapping objects and automatically executes all the method calls in
native OS threads, while the latter needs to be applied to each
method call in the code repeatedly.

Existing calls are modified for the sake of consistency.

Closes-Bug: #1607461

(cherry picked from commit 3405a28688eacbca23cf5cac0a611d33fb1a1f2c)

Change-Id: I5ce75649ec6dec9f705ef1d8d14d73f293233ffb
Signed-off-by: zhangjianfeng <kill51216@163.com>