[bugfixes-backport]Refresh info_cache after deleting floating IP
When deleting a floating IP associated with Neutron's info_cache we don't
refresh the info_cache after it is deleted.
This patch makes it so the info_cache is refreshed when an associated
floating IP is deleted. If there is no info_cache associated with the
floating IP then info_cache is not refreshed.

Closes-Bug:  #1614538
(cherry picked from commit cdb9b6820dc17971bca24adfc0b56f030f0ae827)

Change-Id: I6d04223b7a66ddd28b53e769e0ddbc5588a5b016
Signed-off-by: zhangjianfeng <kill51216@163.com>