[bugfixes-backport]refresh instances_path when shared storage used
When doing Live migration with shared storage, it happens erratically,
that the check for the shared storage test_file fails. Because the shared
volume is under heavy IO (many instances on many compute nodes) the client
does not immediately sees the new content of the folder. This delay
could take up to 30s.
This can be fixed if the client is forced to refresh the directories
content, which can be achieved by 'touch' on the directory. Doing so,
the test_file is visibile instantly, within ms.
The patch adds a 'touch' on instances_path in check_shared_storage_test_file,
before checking the existence of the file.

Conflicts:
    nova/tests/unit/virt/libvirt/test_driver.py

NOTE(lyarwood): Conflict caused by the signature of
_check_shared_storage_test_file changing as part of I6cca257

Closes-Bug: #1617299
(cherry picked from commit 1af73d1fb3169c5b3cce77d94316922496bbaf9a)

Change-Id: Ief42d2974a23b506d4692d87dbb475cebfefbdcc
Signed-off-by: zhangjianfeng <kill51216@163.com>