[bugfixes-backport]virt: handle unicode when logging LifecycleEvents
The repr on the LifecycleEvent object includes a translated
name, which blows up with a UnicodeEncodeError in the emit_event
method because of str(event) and non-English locales.

This change uses six.text_type on the object rather than str
and adds a test to recreate the bug and show the fix.

Closes-Bug: #1621392
(cherry picked from commit 2b57b3d867f568a1539d4419b441ae8d666b77c4)

Change-Id: Ib3047808f7a02fa18eb0c857420d62d071175f50
Signed-off-by: zhangjianfeng <kill51216@163.com>