[bugfixes-backport]Fix error status code on update-volume API
As the following part of API-WG guidline[1],

  If a request contains a reference to a nonexistent resource in the
  body (not URI), the code should be 400 Bad Request. Do not use 404
  NotFound because :rfc:`7231#section-6.5.4` (section 6.5.4) mentions
  the origin server did not find a current representation for the
  target resource for 404 and representation for the target resource
  means a URI

Nova should return a BadRequest response(400) in this case, because
new_volume_id is specified in a request body. old_volume_id is not
necessary to be changed because the value is specified with URI.
So it is valid to return NotFound response if that is not existent.

[1]: https://github.com/openstack/api-wg/blob/master/guidelines/http.rst#failure-code-clarifications

Close-Bug: #1629110
(cherry picked from commit edd86d9dac1ea75bc580a7964e7d699ee9644b19)

Change-Id: Ic7ead633fecb86d2bf28c5b1acbc4240e62c2fb3
Signed-off-by: zhangjianfeng <kill51216@163.com>