[bugfixes-backport]Changed an HTTP exception to return proper code
POSTing to /servers with a content-type of text/plain
and a text/plain body results in a response code of 400.
It should be 415. I found this line in the code that
appears to handle this singular case and modified
the HTTP exception used to the correct one. Tests were
also updated accordingly.

bug from:https://github.com/openstack/nova/commit/d627ae852618fff9eebfc27fdf73287e45b1a7c7

Change-Id: Ic021feb62ccb7b70dd2f886e6525e581a84512ad