[bugfixes-backport]<VMWare> Allocate free bus for new SCSI controller
There are at most 16 virtual disks that can be attached onto a SCSI
controller, while if the limitation is reached, error occurs
with message VMwareDriverException: Invalid configuration for
device '0'.

The reason is that when the controller's quota is used up, a new
controller will be created automatically, but the newly-created
one will try to use bus number 0, which is already used my existing
controller.

This change will allocate a new free bus number for the newly-created
controller.

Closes bug: #1522232

Change-Id: I83d02a8e600624b9e526f9cd818339f4922ebace