[bugfixes-backport]Only reset dns_name when unbinding port if DNS is  integrated
We only set the dns_name on the port if the DNS Integration
extension is enabled, so we should only reset it on the port
when unbinding if the extension is enabled, otherwise some
Neutron stadium plugins will fail since they don't expect
the 'dns_name' attribute to be on the port.

NOTE(mriedem): This backport adds an import that wasn't in
stable/mitaka but is used in the unit test from the original
patch.

Closes-Bug: #1574565
(cherry picked from commit b256cae8e204fbbf6f3d40f5f4d47013be018a6d)

Change-Id: I78255208e676d8760ba7de9f8352732b6578ebad