libvirt: release serial console ports when destroying  guests
The part of code responsible to release serial ports does not cover
all the cases. This commit moves that part inside the method _destroy
which is responsible for destroying the guest from a libvirt
perspective which is when we want to release ports.

(cherry picked from commit 95da4e87881424524c8ddda6a546764d43eb755e)

Change-Id: Ibd4b262cf66152e08020218898847767ad2e08cd
Closes-Bug: #1489853